Second version of taxonomy CRUD
I need to develop the process of editing an existing taxonomy in the database. It is hard because the paper_taxonomy table has a foreign key to the taxonomy fields. So, I need to check if the field is related to a paper and only allow the deletion when such foreign key does not exist.